requirements.txt
-i https://pypi.org/simple
asgiref==3.11.0; python_version >= '3.9'
django==4.2.27; python_version >= '3.8'
django-cors-headers==4.9.0; python_version >= '3.9'
djangorestframework==3.16.1; python_version >= '3.9'
sqlparse==0.5.5; python_version >= '3.8'
typing-extensions==4.15.0; python_version >= '3.9'
tzdata==2025.3; python_version >= '2'
